requirements.txt
django==4.2.17
python-decouple==3.8
django-model-utils==5.0.0
django_admin_interface==0.29.3
drf-spectacular==0.28.0
gunicorn==23.0.0
django-modeltranslation==0.19.12
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.4.0
django-phonenumber-field==8.0.0
phonenumbers==8.13.53
pyotp==2.9.0
eskiz_sms==0.2.3
django-model-utils==5.0.0
django_filter==24.3
django-mptt==0.16.0
